Bug Fix 2: "GPIO.remove_event_detect"
Bug Fix 2: The "GPIO.add_event_detect" function is used for the GPIO.BCM input 5 and 6. These processes then run in the background parallel to the main program. These processes must also be stopped again if the program is ended with "except KeyboardInterrupt" , otherwise there will be overlaps and the bouncing phenomenon during a soft start.